Rename source files.
